requirements.txt
snowflake-connector-python==3.0.*